# Improve ping and pong performance