--- a/lib/Sender.js
+++ b/lib/Sender.js
@@ -11,6 +11,12 @@ const ErrorCodes = require('./ErrorCodes');
 const bufferUtil = require('./BufferUtil').BufferUtil;
 const PerMessageDeflate = require('./PerMessageDeflate');
 
+const emptyPing       = new Buffer([0x89, 0x00]);
+const emptyPingMasked = new Buffer([0x89, 0x80, 0x00, 0x00, 0x00, 0x00]);
+
+const emptyPong       = new Buffer([0x8a, 0x00]);
+const emptyPongMasked = new Buffer([0x8a, 0x80, 0x00, 0x00, 0x00, 0x00]);
+
 /**
  * HyBi Sender implementation, Inherits from EventEmitter.
  */
@@ -57,12 +63,11 @@ class Sender extends EventEmitter {
    */
   ping(data, options) {
     var mask = options && options.mask;
-    var self = this;
-    this.messageHandlers.push(function(callback) {
-      self.frameAndSend(0x9, data || '', true, mask);
-      callback();
-    });
-    this.flush();
+    if (data) {
+      this.frameAndSend(0x9, data, true, mask);
+    } else {
+      sendFramedData.call(this, mask ? emptyPingMasked : emptyPing);
+    }
   }
 
   /**
@@ -72,13 +77,11 @@ class Sender extends EventEmitter {
    */
   pong(data, options) {
     var mask = options && options.mask;
-    var self = this;
-    this.messageHandlers.push(function(callback) {
-      self.frameAndSend(0xa, data || '', true, mask);
-      callback();
-    });
-
-    this.flush();
+    if (data) {
+      this.frameAndSend(0xa, data, true, mask);
+    } else {
+      sendFramedData.call(this, mask ? emptyPongMasked : emptyPong);
+    }
   }
 
   /**