Skip messageHandler mechanism for ping/pong packets and use premade empty ping and pong frames.

Improves the included ping benchmark from

```
ping message x 432,273 ops/sec ±1.38% (86 runs sampled)
ping with no data x 706,276 ops/sec ±1.79% (85 runs sampled)
```

to

```
ping message x 662,729 ops/sec ±0.77% (87 runs sampled)
ping with no data x 18,062,074 ops/sec ±1.18% (85 runs sampled)
```