@lpinca 2016-09-29
I think that skipping the messageHandler mechanism might change the order of messages.