@lpinca 2016-10-02
Order is indeed changed. Here is a test case:

```js
'use strict';

const ws = require('.');

const server = new ws.Server({
  clientTracking: false,
  port: 8080
}, () => {
  const socket = new ws('ws://localhost:8080');

  socket.on('message', (message) => console.log(message));
  socket.on('ping', () => console.log('ping'));
  socket.on('close', () => server.close());
});

server.on('connection', (socket) => {
  socket.send('hello');
  socket.ping();
  socket.close();
});
```

```shell
$ node gh-848.js 
ping
hello
```

With this said, the overhead added by the messageHandler mechanism is significant and it would be nice to have it turned off when permessage-deflate is disabled.