@Nibbler999 2016-10-03
Does that matter? I don't think the websockets spec says that order needs to be preserved between ping and data frames. I'm sure not all implementations do so.