@lpinca 2016-10-03
Well one of the nice features of the WebSocket protocol is that the order of messages is preserved. I'm not sure if the order should be respected with ping/pong frames but I would expect that.